LayoutTests/fast/dom/Document/scrollingElement-standards-mode.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Tests the behavior of document.scrollingElement in standards mode.");

debug("In standards mode the scrolling element is supposed to be `HTML`");
debug("INTENTIONAL FAILURE. This can be fixed once https://bugs.webkit.org/show_bug.cgi?id=106133 is fixed.");
shouldBe("document.scrollingElement", "document.documentElement");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
