LayoutTests/fast/dom/Document/scrollingElement-standards-mode-expected.txt

Tests the behavior of document.scrollingElement in standards mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


In standards mode the scrolling element is supposed to be `HTML`
INTENTIONAL FAILURE. This can be fixed once https://bugs.webkit.org/show_bug.cgi?id=106133 is fixed.
FAIL document.scrollingElement should be [object HTMLHtmlElement]. Was [object HTMLBodyElement].
PASS successfullyParsed is true

TEST COMPLETE
